--- fastfield_codecs/src/bitpacked/mod.rs
@@ -1,5 +1,0 @@
-mod reader;
-mod serialize;
-
-pub use reader::BitpackedFastFieldReader;
-pub use serialize::BitpackedFastFieldSerializer;
